fix: extract common logic
