+++ apps/journeys-admin/pages/index.tsx
@@ -6,20 +6,16 @@ import {
   withAuthUserTokenSSR
 } from 'next-firebase-auth'
 import { NextSeo } from 'next-seo'
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useTranslation } from 'react-i18next'
-import { getLaunchDarklyClient } from '@core/shared/ui/getLaunchDarklyClient'
 import { gql } from '@apollo/client'
 import { useFlags } from '@core/shared/ui/FlagsProvider'
 import { UserInviteAcceptAll } from '../__generated__/UserInviteAcceptAll'
 import { JourneyList } from '../src/components/JourneyList'
 import { PageWrapper } from '../src/components/NewPageWrapper'
-import { createApolloClient } from '../src/libs/apolloClient'
-import i18nConfig from '../next-i18next.config'
 import { OnboardingPanelContent } from '../src/components/OnboardingPanelContent'
 import { TeamSelect } from '../src/components/Team/TeamSelect'
 import { TeamMenu } from '../src/components/Team/TeamMenu'
-import { checkConditionalRedirect } from '../src/libs/checkConditionalRedirect'
+import { initAndAuthApp } from '../src/libs/initAndAuthApp'
 
 export const ACCEPT_USER_INVITE = gql`
   mutation UserInviteAcceptAll {
@@ -54,29 +50,11 @@ function IndexPage(): ReactElement {
 export const getServerSideProps = withAuthUserTokenSSR({
   whenUnauthed: AuthAction.REDIRECT_TO_LOGIN
 })(async ({ AuthUser, locale }) => {
-  const ldUser = {
-    key: AuthUser.id as string,
-    firstName: AuthUser.displayName ?? undefined,
-    email: AuthUser.email ?? undefined
-  }
+  const { apolloClient, flags, redirect, translations } = await initAndAuthApp({
+    AuthUser,
+    locale
+  })
 
-  // run independent tasks (getting LaunchDarkly client, user's ID token, and server-side translations) concurrently
-  const [translations, launchDarklyClient, token] = await Promise.all([
-    serverSideTranslations(
-      locale ?? 'en',
-      ['apps-journeys-admin', 'libs-journeys-ui'],
-      i18nConfig
-    ),
-    getLaunchDarklyClient(ldUser),
-    AuthUser.getIdToken()
-  ])
-
-  const flags = (await launchDarklyClient.allFlagsState(ldUser)).toJSON() as {
-    [key: string]: boolean | undefined
-  }
-
-  const apolloClient = createApolloClient(token != null ? token : '')
-  const redirect = await checkConditionalRedirect(apolloClient, flags)
   if (redirect != null) return { redirect }
 
   await apolloClient.mutate<UserInviteAcceptAll>({
