+++ apps/journeys-admin/src/libs/initAndAuthApp/index.ts
@@ -1,0 +1,1 @@
+export { initAndAuthApp } from './initAndAuthApp'
